Reduced Cost, and expand supported types by replacing C++ functions with Macros
Code module removed.
Now this is a content plugin.
Macro Library holds two macros
- Default
- Default Scale
The new Default Macro covers all allowed types via wildcard, which supports structs too.
Also benefit to this is that it is no longer a function call to get a hardcoded value. Macro is collapsed into the blueprint. This should add virtually no cost, and be cheaper than the function call.